First Iteration WeEcosystem
As I could not upload the unzipped folder containing WeEocystem through the GitHub website, through Greenfoot I exported the Scenario as a standalone project gfar file. I am not sure if this will work or not as it is past 12:00 AM because it took us a little longer to combine our code than we thought.